requirements.txt
PyYAML==6.0.1
Logbook==1.7.0.post0
wxPython==4.2.1